+++ web/routes.py
@@ -6,6 +6,7 @@ from flask import render_template, Flask,request
 from web.database import workdir
 from web import app
 from web.routes_util import containers, overview, system_info
+from web.routes_util.containers import docker_client, container_detail, removed_containers
 
 @app.route('/health')
 def health():
@@ -28,6 +29,16 @@ def internal_server_error(error):
 
 @app.route('/test')
 def test():
+    params = container_detail()
     return render_template(
-        "test.html"
+        "test.html",
+        paused=params[0],
+        data=params[1],
+        running=params[2],
+        restarting=params[3],
+        stopped=params[4],
+        states=params[5],
+        dockerweb = params[6],
+        removed_containers=removed_containers,
+        section='containers'
     )